move full rebuild process into multithreading package
